pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["jailbreakerVC <54243183+jailbreakerVC@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3.2"
bardapi = "^0.1.11"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.88.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
